+++ regenerate-compilation-failures.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+set -e
+
+TRYBUILD=overwrite cargo test