+++ src/commands/control_flow_commands.rs
@@ -5,12 +5,11 @@ pub(crate) struct IfCommand {
     condition: NextItem,
     true_code: InterpretationStream,
     false_code: Option<InterpretationStream>,
+    nothing_span_range: SpanRange,
 }
 
 impl CommandDefinition for IfCommand {
     const COMMAND_NAME: &'static str = "if";
-
-    const OUTPUT_BEHAVIOUR: CommandOutputBehaviour = CommandOutputBehaviour::AppendStream;
 
     fn parse(mut arguments: InterpreterParseStream) -> Result<Self> {
         static ERROR: &str = "Expected [!if! (condition) { true_code }] or [!if! (condition) { true_code } !else! { false_code}]";
@@ -37,12 +36,13 @@ impl CommandDefinition for IfCommand {
             condition,
             true_code,
             false_code,
+            nothing_span_range: arguments.full_span_range(),
         })
     }
 }
 
 impl CommandInvocation for IfCommand {
-    fn execute(self: Box<Self>, interpreter: &mut Interpreter) -> Result<InterpretedStream> {
+    fn execute(self: Box<Self>, interpreter: &mut Interpreter) -> Result<CommandOutput> {
         let evaluated_condition = self
             .condition
             .interpret_as_expression(interpreter)?
@@ -50,13 +50,15 @@ impl CommandInvocation for IfCommand {
             .expect_bool("An if condition must evaluate to a boolean")?
             .value();
 
-        if evaluated_condition {
-            self.true_code.interpret_as_tokens(interpreter)
+        let output = if evaluated_condition {
+            self.true_code.interpret_as_tokens(interpreter)?
         } else if let Some(false_code) = self.false_code {
-            false_code.interpret_as_tokens(interpreter)
+            false_code.interpret_as_tokens(interpreter)?
         } else {
-            Ok(InterpretedStream::new())
-        }
+            InterpretedStream::new(self.nothing_span_range)
+        };
+
+        Ok(CommandOutput::AppendStream(output))
     }
 }
 
@@ -64,12 +66,11 @@ impl CommandInvocation for IfCommand {
 pub(crate) struct WhileCommand {
     condition: NextItem,
     loop_code: InterpretationStream,
+    nothing_span_range: SpanRange,
 }
 
 impl CommandDefinition for WhileCommand {
     const COMMAND_NAME: &'static str = "while";
-
-    const OUTPUT_BEHAVIOUR: CommandOutputBehaviour = CommandOutputBehaviour::AppendStream;
 
     fn parse(mut arguments: InterpreterParseStream) -> Result<Self> {
         static ERROR: &str = "Expected [!while! (condition) { code }]";
@@ -83,13 +84,14 @@ impl CommandDefinition for WhileCommand {
         Ok(Self {
             condition,
             loop_code,
+            nothing_span_range: arguments.full_span_range(),
         })
     }
 }
 
 impl CommandInvocation for WhileCommand {
-    fn execute(self: Box<Self>, interpreter: &mut Interpreter) -> Result<InterpretedStream> {
-        let mut output = InterpretedStream::new();
+    fn execute(self: Box<Self>, interpreter: &mut Interpreter) -> Result<CommandOutput> {
+        let mut output = InterpretedStream::new(self.nothing_span_range);
         let mut iteration_count = 0;
         loop {
             let evaluated_condition = self
@@ -113,6 +115,6 @@ impl CommandInvocation for WhileCommand {
                 .interpret_as_tokens_into(interpreter, &mut output)?;
         }
 
-        Ok(output)
+        Ok(CommandOutput::AppendStream(output))
     }
 }