+++ src/interpretation/interpreted_stream.rs
@@ -2,26 +2,23 @@ use crate::internal_prelude::*;
 
 #[derive(Clone)]
 pub(crate) struct InterpretedStream {
+    source_span_range: SpanRange,
     token_stream: TokenStream,
 }
 
 impl InterpretedStream {
-    pub(crate) fn new() -> Self {
+    pub(crate) fn new(source_span_range: SpanRange) -> Self {
         Self {
+            source_span_range,
             token_stream: TokenStream::new(),
         }
     }
 
-    pub(crate) fn raw(token_stream: TokenStream) -> Self {
-        Self { token_stream }
-    }
-
-    pub(crate) fn of_ident(ident: Ident) -> Self {
-        TokenTree::Ident(ident).into()
-    }
-
-    pub(crate) fn of_literal(literal: Literal) -> Self {
-        TokenTree::Literal(literal).into()
+    pub(crate) fn raw(empty_span_range: SpanRange, token_stream: TokenStream) -> Self {
+        Self {
+            source_span_range: empty_span_range,
+            token_stream,
+        }
     }
 
     pub(crate) fn extend(&mut self, interpreted_stream: InterpretedStream) {
@@ -61,6 +58,18 @@ impl InterpretedStream {
         self.token_stream.is_empty()
     }
 
+    pub(crate) fn as_singleton(self, error_message: &str) -> Result<TokenTree> {
+        if self.is_empty() {
+            return self.source_span_range.err(error_message);
+        }
+        let mut iter = self.token_stream.into_iter();
+        let first = iter.next().unwrap(); // No panic because we're not empty
+        match iter.next() {
+            Some(_) => self.source_span_range.err(error_message),
+            None => Ok(first),
+        }
+    }
+
     pub(crate) fn into_token_stream(self) -> TokenStream {
         self.token_stream
     }
@@ -69,6 +78,7 @@ impl InterpretedStream {
 impl From<TokenTree> for InterpretedStream {
     fn from(value: TokenTree) -> Self {
         InterpretedStream {
+            source_span_range: value.span_range(),
             token_stream: value.into(),
         }
     }
@@ -76,6 +86,10 @@ impl From<TokenTree> for InterpretedStream {
 
 impl HasSpanRange for InterpretedStream {
     fn span_range(&self) -> SpanRange {
-        self.token_stream.span_range()
+        if self.token_stream.is_empty() {
+            self.source_span_range
+        } else {
+            self.token_stream.span_range()
+        }
     }
 }