+++ tests/compilation_failures/expressions/compare_int_and_float.rs
@@ -1,0 +1,7 @@
+use preinterpret::*;
+
+fn main() {
+    let _ = preinterpret!{
+        [!evaluate! 5 < 6.4]
+    };
+}