+++ tests/compilation_failures/expressions/fix_me_comparison_operators_are_not_lazy.rs
@@ -1,0 +1,11 @@
+use preinterpret::*;
+
+fn main() {
+    preinterpret!{
+        [!set! #is_eager = false]
+        let _ = [!evaluate! false && ([!set! #is_eager = true] true)];
+        [!if! #is_eager {
+            [!error! "The && expression is not evaluated lazily" []]
+        }]
+    }
+}