+++ tests/compilation_failures/expressions/fix_me_negative_max_int_fails.rs
@@ -1,0 +1,12 @@
+use preinterpret::*;
+
+fn main() {
+    let _ = preinterpret! {
+        // This should not fail, and should be fixed.
+        // This test just records the fact it doesn't work as a known issue.
+        // A fix of this should remove this test and move it to a working test.
+        [!evaluate! -128i8]
+        // This should also not fail according to the rules of rustc.
+        [!evaluate! -(--128i8)]
+    };
+}