+++ tests/simple_test.rs
@@ -3,6 +3,7 @@ use preinterpret::preinterpret;
 preinterpret! {
     [!set! #bytes = 32]
     [!set! #postfix = Hello World #bytes]
+    [!set! #some_symbols = and some symbols such as [!raw! #] and #123]
     [!set! #MyRawVar = [!raw! Test no #str [!ident! replacement]]]
     [!ignore! non - sensical !code :D - ignored (!)]
     struct MyStruct;